Renamed files to match PEP 0008 file naming conventions.
